backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.12
python-dotenv==1.0.1
aiofiles==24.1.0
# Phase 3: kaira-evals evaluation engine
google-genai>=1.0.0
openai>=1.0.0
anthropic>=0.40.0
aiohttp>=3.9.0
pandas>=2.0.0
google-auth>=2.0.0
playwright==1.52.0
# Auth
PyJWT>=2.8.0
bcrypt>=4.1.0
